Improvements text
